@@ -10,8 +10,7 @@ import { ObservableService } from '../services/observable.service';
 export class ChildComponent implements OnInit {
 
 
-
-  @Input('msgInput') msg = new String(''); //Decorador Input para poder recibir datos del padre
+  @Input() msg = ''; //Decorador Input para poder recibir datos del padre
   @Output() childMsg = new EventEmitter<string>(); //Decorador del Output para enviar datos al padre
 
 
